[REL] stock_picking_state to V11. [FIX] Separate classes and views by model. [FIX] Minor fix.
