+++ stock_picking_state/models/__init__.py
@@ -2,4 +2,5 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import stock
+from . import stock_picking
+from . import stock_picking_state_detail
