--- stock_picking_state/views/stock_menu.xml
@@ -1,22 +1,0 @@
-<odoo>
-    <data>
-     <record id="stock_picking_state_detail_action" model="ir.actions.act_window">
-            <field name="name">Picking State Detail</field>
-            <field name="res_model">stock.picking.state_detail</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="view_id" ref=""/>
-            <field name="search_view_id" ref=""/>
-            <field name="domain">[]</field>
-        </record>
-
-        <menuitem
-            name="Picking State Detail"
-            id="menu_state_detail"
-            sequence="20"
-            action="stock_picking_state_detail_action"
-            parent="stock.menu_stock_config_settings"
-            />
-
-    </data>
-</odoo>
