+++ stock_picking_state/views/stock_picking_state_detail.xml
@@ -1,0 +1,34 @@
+<odoo>
+    <record id="stock_picking_state_detail_action" model="ir.actions.act_window">
+        <field name="name">Picking State Detail</field>
+        <field name="res_model">stock.picking.state_detail</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_id" ref=""/>
+        <field name="search_view_id" ref=""/>
+        <field name="domain">[]</field>
+    </record>
+
+    <record id="view_picking_detail_form" model="ir.ui.view">
+        <field name="name">stock.picking_detail</field>
+        <field name="model">stock.picking.state_detail</field>
+        <field name="arch" type="xml">
+            <form>
+                <group>
+                    <field name="name"/>
+                    <field name="sequence"/>
+                    <field name="state"/>
+                    <field name="picking_type"/>
+                </group>
+            </form>
+        </field>
+    </record>
+
+    <menuitem
+        name="Picking State Detail"
+        id="menu_state_detail"
+        sequence="20"
+        action="stock_picking_state_detail_action"
+        parent="stock.menu_stock_config_settings"
+        />
+</odoo>
