--- stock_picking_state/views/stock_view.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-
-        <record id="view_vpicktree" model="ir.ui.view">
-            <field name="name">Stock picking tree</field>
-            <field name="model">stock.picking</field>
-            <field name="inherit_id" ref="stock.vpicktree"/>
-            <field name="arch" type="xml">
-                <field name="state" position="after">
-                    <field name="state_detail_id"/>
-                </field> 
-            </field>  
-        </record>
-
-        <record id="view_picking_form" model="ir.ui.view">
-            <field name="name">Stock picking</field>
-            <field name="model">stock.picking</field>
-            <field name="inherit_id" ref="stock.view_picking_form"/>
-            <field name="arch" type="xml">
-                <header position="after">
-                    <header>
-                        <field name="state_detail_id" widget="statusbar" clickable="True" domain="[('state','=', state),'|',('picking_type','=', picking_type_code), ('picking_type','=', False)]"/>
-                    </header>
-                </header>
-            </field>  
-        </record>
-
-
-        <record id="view_picking_detail_form" model="ir.ui.view">
-            <field name="name">stock.picking_detail</field>
-            <field name="model">stock.picking.state_detail</field>
-            <field name="arch" type="xml">
-                <form>
-                    <group>
-                        <field name="name"/>
-                        <field name="sequence"/>
-                        <field name="state"/>
-                        <field name="picking_type"/>
-                    </group>
-                </form> 
-            </field>  
-        </record>
-
-    </data>
-</odoo>
